requirements.txt
python-jobspy>=1.1.80
requests>=2.31.0
